调整client初始化方式

+++ example/server/main.go
@@ -17,7 +17,7 @@ var (
 	password = ""
 )
 
-func add(a, b int) (int, error) {
+func add(a, b int64) (int64, error) {
 
 	return a + b, nil
 }
